make Hypre test run with pytest by default
To preserve code covereage information, the Travis tests are run in two blocks:
- the 'not Hypre' tests which use only internal code
- the 'Hypre' test which uses the external solver
